CAD-323 Add Github workflow for CI/CD
Follow-up to PR #3 to finalise the deploy pipeline.

This is a basic pipeline with a test, build and deploy jobs. However the test step will do little right now until we actually have... tests.

Also on second thought I've manually created an http basic-auth secret so it can be used in the ingress and the deployed app is not left wide open.
This secret is not yet committed to the repo until we setup git-crypt etc. Credentials will be shared in confluence/slack.