Simplify HostEnv class
Renamed `ENV` to `ENV_NAME` as to be more explicit to what it is and also avoid possible name clashing.

Removed the `test?` method as for this (along with `development?`) is best to always use Rails.env, as HostEnv is about named environments and not if we are running in test/development or production mode.