CRIMAP-12 Add remaining steps classes and first step
Ticket: https://dsdmoj.atlassian.net/browse/CRIMAP-12

Add the remaining step classes (mainly superclass that are inherited to implement the individual form objects and steps).

Implement the first step as an example (client has partner yes-no question).

How to test it: after running migration, start the server and visit the following URL:
http://localhost:3000/steps/client/has_partner

Check errors are produced if no option is selected, and also select any of the options to go to a demo page (to be deleted, is just an example of the decision tree).

Note: some of the code is not yet useful or do nothing, we have to evolve some parts of the apps to wire up all this (for example, the "save and come back later" button, session handling, etc.)

I will add some documentation about all this in a separate doc in a future PR once we have this all more polished.