CRIMAP-16 Update to new ingress controller
Ticket https://dsdmoj.atlassian.net/browse/CRIMAP-16

To improve security of the Cloud Platform they have introduced new v1 Ingress Controllers.

Tweaked the existing ingress manifest as per documentation here:
https://user-guide.cloud-platform.service.justice.gov.uk/documentation/other-topics/Switch-ingress-to-v1-ingress-controller.html#switch-ingress-to-the-new-nginx-ingress-controller-v1-2

And slack convo here:
https://mojdt.slack.com/archives/CH6D099DF/p1657794137979789

This would have incur a minor downtime but because this service is not live/production we can do this all in one go.

To not incur downtime there are specific instructions to do it in 2-steps creating a separate ingress controller.