Remove rails-controller-testing gem
We will try to use `request` specs instead of `controller` specs and this gem uses deprecated methods like `render_template` or `assigns`.

I will leave however for now the step specs in the controllers/ directory as I feel it is testing a bit more that just an input and an output and contain some extra logic around sessions, etc.

A negative thing: assets now need to be compiled and post-processed
for the request tests to work as they seem to render the application
layout and thus the govuk_template.html.erb with the asset tags.