CRIMAP-6 Add basic error handling and error pages (#28)
* Add link_button helper method

This is just a helper method to output html markup for links with the appearance of a button.

Can be passed options to customise the attributes but most of the time the defaults are enough.

* Add basic error handling and error pages

Just a few common errors, like invalid session or `ApplicationNotFound` (this one still not really in use until we have the applications dashboard functionality).

The error pages copy is kind of placeholder, generic copy. This will have to revised at some point in the future.