Register new ActiveModel type for multi param dates
Similar to what we already did for the yes-no value objects, we are registering a new ActiveModel Type, named `:multiparam_date` to use in form objects that use dates.

This custom type is only necessary because if we use the already existing type (`ActiveModel::Type::Date`) it will blow up with a `MultiparameterAssignmentErrors` exception when trying to cast any invalid date.

Also, if the date is invalid it wouldn't return back to the view whatever the user entered in the date, which is not great from a user experience point of view, it will all be blank.

Thus, with this custom type, if the date is invalid `Date.valid_date?` (in addition to not blow up) we return the multi param hash so we can validate this in a custom validator with more granularity.

If the date is valid we just return the date object and still run further validations in the validator.