Implement an extensive custom date validator
This date validator performs most of the heavy validation on multi param dates (apart from any additional ones that can be added to the standard Rails validator, like `presence`).

It works under the assumption that we can have a "valid" date object, but we might want additional logic validations on this, for example some dates, like DOB can't be in the future.

Or we can have an invalid date not casted to a date object, thus we have the raw multi param values (3 values for day, month and year) and we can perform granular validation on these so we can present a more accurate error than just "date is invalid".

For example we validate `day` is no greater than 31, or month is not greater than 12, etc. and set appropriate errors.

It can also check for blank values (like having entered only day and month, but not year) etc.

To use it, just add to your attribute the validation:

```ruby
validates :date_of_birth, presence: true, multiparam_date: true
```

If the attribute is not mandatory and can be left blank, remove `presence: true`.

If dates in the future are allowed (by default they are not), use like this:

```ruby
validates :date_of_birth, presence: true, multiparam_date: { allow_future: true }
```