CRIMAP-56 Custom string type to strip blanks
A common thing seen across web applications is copy&paste data, sometimes from a spreadsheet etc. where inadvertently they introduce some extraneous spacing.

These leading/trailing spaces are kept intact on saving to the database, which then can produce some validation errors in downstream services.

To reduce the number of instances where this happens, we can create a custom `:string` type that literally behave like the usual type but performs a strip if the value is a string.

Note: as part of this, I moved the types we've defined so far to the `Type` namespace.