Refactor people relationships and prepare contact details (#36)
To anticipate for when we have `partner`, it makes more sense to have a `Person` superclass with 2 (for now and probably not to change ever) subclasses `Applicant` and `Partner`.

Also to prepare for the contact details, it makes more sense for these to be a `has_one` relationship on each person. However trying to simplify second level relationships, we expose 2 `has_one through` relationships on the `crime_application` so it keeps behaving kind of like a 1st level relation and our form object are kept simple.