CRIMAP-60 Add home address manual entry (#38)
First part of the journey to enter the home address of the applicant (client's home address).

For now just the manual entry. Postcode lookup will be added later on as a separate one or more PRs.

The code is somehow coupled for now to the applicant as we don't have `partner` yet for the MVP, but in the future I anticipate we could make these controllers / form objects somehow work for partner too with not too much effort.

The database models already supports as of PR #36 to have different contact details for applicant and for partner.