CRIMAP-61 Add postcode lookup step (no API yet)
This just prepares the step to, in a follow-up PR integrate with the Ordnance API.

However before that it made sense to do this step as a normal one, because I think we have to plan to make it generic to work for correspondence addresses too, and in the future even with partners.

I will have a think about how best to do that because if we make it work for this view it will also work for the results of the postcode lookup.

Note: no "postcode" validation yet, we can add that later, for now only presence.