CRIMAP-87 Implement address lookup service (#47)
* Add webmock and faraday gems

We will use these gems to stub HTTP requests and to make the API requests to the Ordnance API.

* Add fixtures for the address lookup results

These are used to stub the http request. These are real examples from the API, just cut down to just a few results each as we don't need more.

* Implement the Ordnance survey API request

Very similar (if not identical) to what was done in Civil Apply and in C100, maybe others. The C100 introduced some improvements that I'm keeping here.

* Fix a rubocop offence