CRIMAP-88 Select address from API results
This is a follow-up to PR #47. In the previous PR the backend was implemented and in this PR it is surfaced in the interface for users to select the address from a dropdown.

Most of the code for the retrieval and the update of the address attributes happen in the form object.

There are some refinements left that will be done as part of another PR, for example resetting the previous address if the postcode change, or adding a (UK) postcode validator as the Ordnance Survey API will only work with UK postcodes anyways.