Perform reset of attributes if there are changes
If the postcode changes, we have to reset previously store address as it will perform a new lookup and the user will have to select a new address based on the new postcode.

Similarly, if we have already an address selected, and attributes saved, if the user changes something on that address, it will reset the `lookup_id` attribute meaning that is no longer the same address previously selected from the dropdown.