Add UK postcode validator
Improve the postcode lookup step by ensuring valid UK postcodes are used, as otherwise Ordnance Survey will fail or return no addresses anyways.

It validates against million of postcodes in England, Wales, Scotland, Northern Ireland, the Channel Islands, and the Isle of Man.

I've also checked and it pass for Guernsey and Jersey postcodes too.

Not really related to this validator but I've renamed the Struct method from `address_lines` to `compact_address` to better capture the meaning, and added the city to it because at first glance in the dropdown a provider might want to have the assurance they are looking at results from a city they might expect, particularly important if we are going to skip the address manual entry page after they select an address in the dropdown.