Sort out form builder defaults and fix markup
We had non-standard defaults set for the form builder. These are good defaults when an app is mainly done by many 1 question per page, as it saves having to set again and again the same tag, size, etc.

But in reality this app will not have that many individual questions and where we have them, we can customise accordingly.

Removed the custom defaults and tweaked the existing pages where needed to output proper markup, tags and sizes.

Also added `autocomplete: 'off'` if it was missing.