Implement inquiry mechanism for value objects (#58)
I've probably spent more time that I should on this little QoL thing, but in the end it is kind of nice :-)

Instead of implementing again and again inquiry methods for each of the values on a value-object (although in many cases it is not necessary if they are not inspected later on), now the  superclass `ValueObject` takes care of defining this "question mark" methods to ask a value object about its values.

It iterates through all defined values in the `VALUES` constant declaring methods like `yes?` or `no?` or `home_address?` etc.

Also moved to the superclass the class method `.values`. This makes for leaner value objects, to the point many will only have the `VALUES` constant and nothing else.

Any more complicated inquiry logic can be defined as separated methods in each value object like before.
Tests are also leaner as no need to test in every instance the inquiry methods, the "definition" of these methods is tested in the superclass.