Create AddressFormDecorator and simplify form
This decorator will be used to enhance the form object with some methods that will help us reuse the same address views for any kind of people (applicant or partner) and address type (home or correspondence).

Also moved out of the form object and into the decorator the code that adds to the addresses collection the counter.