Update decision trees
Now we have ready the correspondence address sub-journey, we link it to the main journey, in the contact details step, if the provider selects the "other address" radio option in the correspondence type.

After finishing this correspondence sub-journey we get back to the home (as we don't have yet the following step).