Simplify and rename `BaseRenderer`
We don't have to include the `UrlHelper` module if we don't use the link helpers (like `link_to`) and instead build the link as any other tag with `tag.a ...`

Removed the `.render` class method as it is no longer used, all different components of the task list are first instantiated (.new) and then rendered (#render) in 2 steps which makes them easier to test.

Finally, instead of propagating the view context through all the task list classes, we can include a couple modules that gives as the `tag` (and `safe_join`, etc.) functionality.