Implement TaskList Collection
This is the main class to interact with the task list. A task list is created by instantiating this class, and behaves like a collection of sections, but has methods to inspect individual tasks as well as we need to know how many have been completed or are in progress yet.