Implement individual task status logic
The main paradigm with the task list is that we have many individual tasks, and each of them have their own rules for when is considered in progress, when is considered completed, etc.

We have 10 tasks in total at the moment, but many of those are not going to be implemented in the journey until post-MVP (for example `supporting evidence`).

So in reality we only need to have a task implementation for tasks that will be reachable to the provider.

For example for a task named `client_details`, we have a corresponding `Tasks::ClientDetails`.

For a task named `:capital_assessment` (not in scope for MVP) we don't need to have yet an implementation, and the `Tasks::BaseTask` will be used in lieu.

This BaseTask implements a `not_applicable?` method always to true, so unless there is a specific subclass overriding this method, the task will be marked as `TaskStatus::NOT_APPLICABLE`.

Note this is a first iteration at solving this problem. It is not trivial. At the moment we are looking for specific attributes to see if a task is completed, or in progress. This is a bit fragile so a better approach maybe using the Form Objects themselves might be possible in follow-up PRs.

After all we only have 1 section ready (client details) so we can't see many patterns emerging yet.