Return applications having applicants with name
This improves a bit the dashboard `#index` (list) action so that it only returns applications that have advanced enough to at the very least have an applicant with a name, as otherwise the dashboard either blows app (if there is no applicant record yet) or shows a blank name, no link (if the applicant has no name).

Move the index controller test to a `request` test `spec/requests/dashboard_spec.rb` as it makes a lot of sense for this to be a request test, and much easier to test (although slower I reckon as we are not mocking and we are using real records).