CRIMAP-94 Task list status logic (#67)
* Add `TaskStatus` value object

This holds all the different statuses a task can have, and a helper method to know which ones are considered as `enabled`, meaning the user can interact with tasks that have these statuses.

* Simplify and rename `BaseRenderer`

We don't have to include the `UrlHelper` module if we don't use the link helpers (like `link_to`) and instead build the link as any other tag with `tag.a ...`

Removed the `.render` class method as it is no longer used, all different components of the task list are first instantiated (.new) and then rendered (#render) in 2 steps which makes them easier to test.

Finally, instead of propagating the view context through all the task list classes, we can include a couple modules that gives as the `tag` (and `safe_join`, etc.) functionality.

* Finalise implementation of the task list components

Individual parts of a task list, these being a section, having one or more tasks, each task having a status_tag.

This was partially done in previous PR #63 but now it is completed, and also a bit of a refactor/rename.

* Implement TaskList Collection

This is the main class to interact with the task list. A task list is created by instantiating this class, and behaves like a collection of sections, but has methods to inspect individual tasks as well as we need to know how many have been completed or are in progress yet.

* Implement individual task status logic

The main paradigm with the task list is that we have many individual tasks, and each of them have their own rules for when is considered in progress, when is considered completed, etc.

We have 10 tasks in total at the moment, but many of those are not going to be implemented in the journey until post-MVP (for example `supporting evidence`).

So in reality we only need to have a task implementation for tasks that will be reachable to the provider.

For example for a task named `client_details`, we have a corresponding `Tasks::ClientDetails`.

For a task named `:capital_assessment` (not in scope for MVP) we don't need to have yet an implementation, and the `Tasks::BaseTask` will be used in lieu.

This BaseTask implements a `not_applicable?` method always to true, so unless there is a specific subclass overriding this method, the task will be marked as `TaskStatus::NOT_APPLICABLE`.

Note this is a first iteration at solving this problem. It is not trivial. At the moment we are looking for specific attributes to see if a task is completed, or in progress. This is a bit fragile so a better approach maybe using the Form Objects themselves might be possible in follow-up PRs.

After all we only have 1 section ready (client details) so we can't see many patterns emerging yet.

* Split task list and dashboard locales