Hide secondary button on some steps (#69)
We decided it doesn't make too much sense to show the "Save and come back later" button right from the beginning when the provider didn't enter anything yet.

As we are taking the provider back to the task list, at least we want to wait until past the client details step so they enter first name / last name and there is something to show if they get back to their list of applications.

Made it so we can disable the secondary action in specific steps by passing `secondary: false`. Default it to show secondary action.