Change from using the session to use the params
We are no longer retrieving the current application from the session but instead all steps routes are scoped to a specific `crime_application` record, with its UUID being part of the URL, so we can retrieve it with `params[:id]`.

Implement the `#create` action which just creates a new record and redirect to the `has_partner` step as this step acts as a screener question and must be asked first thing, for each new application.

Update the developer tools in the footer, as these changes introduce some extra complexity that I'd rather not deal with it in this PR to keep it focused.