CRIMAP-117 Skip address manual entry pages
If the provider finds a suitable address when using the postcode lookup, after selecting this address and submitting the form, we “skip“ the manual entry page.

This applies to home address and correspondence address.

We've been showing these pages as a way to test/debug the address lookup but the original designs and user journey was to not show them.

Still providers can access these pages when entering addresses manually, and will also be accessible from the CYA when we have it.