Some minor fixes to markup and back links (#76)
Mainly changes in the flash banner markup according to the design system guidelines (position above the page h1, avoid using h3 in banner content).

Added the banner to the `govuk-grid-column-full` container also as per guidelines.

Fixed the back link in the task list and removed the back link in the dashboard index as it is not accurate where it will take the user. Civil Apply doesn't have back link in their dashboard either.