Add right hand column to task list, markup tidy up
Last batch (maybe) of refinements, tidy up and TLC.

Added the right hand column to the task list page, showing the LAA reference. We pick this from the presenter.

Made the reference in the presenter be a bit more dynamic so is more evident clicking through different applications in the dashboard takes you to different task lists (I think the task list should include also the client's name, I've added a comment in figma design).

As we now use the presenter in the task list, and in the destroy page, made the code in the controller a bit more generic.

Tidy up some html markup and extracted to locales some hardcoded strings.