Add client details to task list aside (#84)
I totally missed there was another figma slide where these details are shown.

Glad we now have the presenter all setup so it was kind of just reusing almost everything, I just added a couple more methods to the presenter.

DRY the time/date format. The `datetime` format is not used yet, but will be used for the stamp date.