CRIMAP-109 Initial co-defendants DB schema and model (#87)
CRIMAP-109 Initial codefendants DB schema and model

Just the basic schema that we will require for the co-defendants (a case has many co-defendants, a co-defendant belongs to a case).

Added the routes as well. Rename the existing `cases` namespace to be singular `case` to follow the pattern we have (like `client` or `address`).