CRIMAP-109 Enter codefendant names (#89)
* CRIMAP-109 Enter codefendant names

Follow-up to previous preparation PR #87.

This implements the controller and form object(s) needed for the CRUD step to add/edit/delete co-defendants.

Note: in this PR only add and edit actions are implemented. In a future PR the delete will also be implemented.

Also only first_name and last_name attributes in this PR. Conflict of interest is a separate ticket and will be added later, otherwise the PR grows too much.

There are some ancillary changes in this PR that were needed as this was our first CRUD step really (addresses was kind of different).

The way this works is slightly different to other steps. Basically, we have a main form object, named `CodefendantsForm` which is in charge of rendering the view which is composed of a collection of `CodefendantFieldsetForm` which are like mini form objects representing each of the individual co-defendant records.

The validation is performed in these individual co-defendant form objects but propagated to the parent form `CodefendantsForm` so errors can be rendered in the summary as well, and are indexed pointing to the specific field with error (as there may be more than 1 co-defendant and all their attributes are named the same really, we have to index them).

* Use the alias from the module HasOneAssociation

As the alias is defined in the `Steps::HasOneAssociation` module we can remove individual alias in form objects.