CRIMAP-110 Add conflict of interest radio options
Another follow-up to all previous PRs related to co-defendants.

This I think is the last one. We add the conflict of interest yes-no radios.

Pretty simple PR as almost everything has been done already in previous PRs.

Only pain was configuring the `govuk_collection_radio_buttons`, because we are using nested attributes, the locales are not automagically fetched for us and we must pass them explicitly to the legend and to both options (the yes and the no).

I've raised an issue in the `govuk-formbuilder` gem repository for a possible enhancement that would make these nested attribute locales work automagically too so we could simplify a lot the code in this view if that happens.