Add an ERB linter and initial support for custom cops
This is more of a learning exercise but I went through the rabbit hole of learning (barely) how to code custom Rubocop cops and when I got the matcher working I realised Rubocop doesn't run on ERB files.

So I researched a bit and found out a suitable ERB linter made by the nice folks at Shopify which lets you, in addition to its own linting (including some HTML linting), runs Rubocop on the ERB "ruby" code.

As expected there are some offences already in the views we already have, that's why I've disabled for now some linters.

In a separate PR I will enable some of them and fix the offences, mostly related to simple things like indentation, spacing, double quotes instead of single quotes, etc. Some of these may be annoying so we can discuss how comfortable we want to be LOL.

NOTE: the erb linter will now run as part of the other linters and tests we have, and also on CI.