Enable case details in task list
Now that we have some of the case details steps, we can update this task so it links to the first step (URN) and also to know when it is in progress or not.

We will not be able to set the "completed" status tag until we have finalised all the case details step so that will need to change sometime in the future.