Cleanup temporary code from co-defendants form object
As we now have the previous yes-no step, I'm removing some of the temporary code I added in the co-defendants form object (and cleaned up the specs).

The temporary code was introduced to simplify development and testing, by creating a blank co-defendant record to start with, just by visiting the co-defendants page.
The decision tree is now in charge of handling this logic as it is bad practise to create records on GET requests.

I've also taken care of resetting any previously entered co-defendants, if the provider go back to the `yes-no` question and selects `no`.