Enable Rubocop in ERB linter and fix offenses
As a first step, I'm enabling Rubocop in the available ERB linters.

This will run Rubocop on any ruby code found in ERB templates, with the same configuration we have for the rest of the code base. This includes our custom cops (for now just the autocomplete, as per PR #94).

However some cops are highly annoying or return false positives (i.e. Style/BlockDelimiters), thus we can tweak the configuration to disable some of these cops on ERB code, while still using them for normal non-ERB code, to get us to a good balance.

I had to fix a few very minor offences. I did individual commits with each of the fixes for better context.