Replace `yes_no` custom attribute with generic one
The `yes_no` custom attribute is a QoL or syntax sugar addition to facilitate automatic casting of the commonly used yes-no radios.

This makes for simpler form objects, with less boilerplate code and more granted for free. For example there is no need to use a `#string_choices` method to validate inclusion, etc.

With these changes, we are removing the `yes_no` attribute but adding a new `value_object` attribute. This new attribute is generic and can be passed the "source" class of the values. All the rest behaves the same.