Remove override of `type_cast_for_schema`
It is not called no matter what I try so I think is safe to just go with whatever does the superclass and if we ever detect any issues we can always implement it.

The documentation is also misleading as in theory this is suppose to be a private method but currently it is public.