Improve handling of value objects in form attributes (#98)
* Replace `yes_no` custom attribute with generic one

The `yes_no` custom attribute is a QoL or syntax sugar addition to facilitate automatic casting of the commonly used yes-no radios.

This makes for simpler form objects, with less boilerplate code and more granted for free. For example there is no need to use a `#string_choices` method to validate inclusion, etc.

With these changes, we are removing the `yes_no` attribute but adding a new `value_object` attribute. This new attribute is generic and can be passed the "source" class of the values. All the rest behaves the same.

* Use the new `value_object` attribute in forms

Replace previous `yes_no` attribute with new generic `value_object` attribute.

Use the `value_object` attribute in forms where any of their attributes are value objects, like `CaseType` or `CorrespondenceType`.

* Update decision tree

As the attribute is already a value object instance, we don't need to do this anymore.

* Remove override of `type_cast_for_schema`

It is not called no matter what I try so I think is safe to just go with whatever does the superclass and if we ever detect any issues we can always implement it.

The documentation is also misleading as in theory this is suppose to be a private method but currently it is public.