CRIMAP-134 Initial offences DB schema and models
As we start to tackle the offences step, we need a series of DB tables and models to represent the data and relationships.

This might change slightly later on if we find out some missing attribute, or weird thing, but overall is the high level representation of what we want.

In a later PR a DB seed migration will be created to populate the `Offences` table based on the spreadsheet.