CRIMAP-125 Initial commit of offences CSV
Store the CSV file containing the pre-filtered list of offences and their attributes, and a thin layer to read it and perform basic operations like find by code (or by name if that make sense), and retrieve all offences.

This is inspired by HMCTS common platform with a very similar approach:
https://github.com/ministryofjustice/hmcts_common_platform

I've made some small tweaks and modifications to better fit our purposes (also implemented the equality methods), but overall is exactly the same.

We might not need the Offences DB table anymore but for now I'm not doing anything on that front until we've validated this other alternative (CSV file).

I've just reused the same `Offence` model but removed its ActiveRecord capabilities so it is just a simple PORO/entity for now.