CRIMAP-125 Initial setup of `Charges` step
Routes, controller, form object and view with locales for a very basic initial boilerplate of the `Charges` step.

This already takes care of creating a `charge` record to work with (CRUD) but it only has an attribute, `offence_name` and no validation etc.

This is so we have something to start with and there are not too much code conflicts/clashing.