CRIMAP-127 Implement offences summary page
This is as well kind of the boilerplate (although it is totally functional) for the offences summary page, where offences will show and they can be changed, removed, or add another.

For now it only shows the yes-no radios to add another offence, with validation, and the decision tree has been updated to introduce this new flow.

In a follow-up PR, I will add to the summary page the list of offences, as well as the additional edit/delete actions.